label-studio/backends/layoutlmv3_backend/requirements.txt
transformers
torch
torchvision
pillow
opencv-python
pytesseract